refactoring and naver map move schedule add screen
