+++ lib/core/router/app_router.dart
@@ -4,7 +4,7 @@ import 'package:dutytable/features/calendar/presentation/views/personal/personal
 import 'package:dutytable/features/calendar/presentation/views/shared/shared_calendar_screen.dart';
 import 'package:dutytable/features/notification/presentation/views/notification_screen.dart';
 import 'package:dutytable/features/profile/presentation/views/profile_screen.dart';
-import 'package:dutytable/features/schedule/presentation/views/schedule_add_screen.dart';
+import 'package:dutytable/features/schedule/presentation/views/add/schedule_add_screen.dart';
 import 'package:dutytable/features/schedule/presentation/views/detail/schedule_detail_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
