--- lib/features/schedule/data/models/address_search_result_model.dart
@@ -1,10 +1,0 @@
-class AddressSearchResult {
-  final String title;
-  final String address;
-
-  AddressSearchResult({required this.title, required this.address});
-
-  factory AddressSearchResult.fromJson(Map<String, dynamic> json) {
-    return AddressSearchResult(title: json['title'], address: json['address']);
-  }
-}
