+++ lib/features/schedule/presentation/views/add/widgets/done_status_section.dart
@@ -1,0 +1,48 @@
+import 'package:dutytable/core/configs/app_colors.dart';
+import 'package:dutytable/features/schedule/presentation/viewmodels/schedule_add_view_model.dart';
+import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
+
+class DoneStatusSection extends StatelessWidget {
+  /// 완료 상태
+  const DoneStatusSection({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    final viewModel = context.watch<ScheduleAddViewModel>();
+
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        Text(
+          "일정 완료 상태",
+          style: TextStyle(fontSize: 16.0, fontWeight: FontWeight.w800),
+        ),
+
+        const SizedBox(height: 10),
+
+        Container(
+          padding: const EdgeInsets.all(10.0),
+          decoration: BoxDecoration(
+            borderRadius: BorderRadius.circular(10),
+            border: Border.all(color: AppColors.cardBorder(context), width: 2),
+          ),
+          child: Row(
+            mainAxisAlignment: MainAxisAlignment.spaceBetween,
+            children: [
+              Text(
+                "일정 완료 상태",
+                style: TextStyle(fontSize: 16.0, fontWeight: FontWeight.w800),
+              ),
+
+              Switch(
+                value: viewModel.isDone,
+                onChanged: (value) => viewModel.isDone = value,
+              ),
+            ],
+          ),
+        ),
+      ],
+    );
+  }
+}
