+++ lib/features/schedule/presentation/views/add/widgets/location_section.dart
@@ -1,26 +1,22 @@
 import 'package:dutytable/core/configs/app_colors.dart';
 import 'package:dutytable/features/schedule/presentation/viewmodels/schedule_add_view_model.dart';
+import 'package:dutytable/features/schedule/presentation/views/location_search_screen.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_naver_map/flutter_naver_map.dart';
 import 'package:provider/provider.dart';
 
-class AddressSection extends StatelessWidget {
-  /// 주소 - 선택 사항
-  /// 주소입력 버튼 클릭 - 주소, 이미지 없을 때
-  /// 주소 입력 후 완료 하면 버튼 UI 바뀜(주소랑 + 지도 이미지)
-  /// 다시 누르면 다시 주소입력 화면으로 이동
-  /// 주소 + 지도 이미지 - 주소, 이미지 있을 때
-  /// X버튼으로 지우면 원래 버튼 UI로 복귀
-  const AddressSection({super.key});
+class LocationSection extends StatelessWidget {
+  const LocationSection({super.key});
 
   @override
   Widget build(BuildContext context) {
-    final vm = context.watch<ScheduleAddViewModel>();
+    final viewModel = context.watch<ScheduleAddViewModel>();
 
     return Column(
       crossAxisAlignment: CrossAxisAlignment.start,
       children: [
         const Text(
-          "주소",
+          "일정 장소",
           style: TextStyle(fontSize: 16, fontWeight: FontWeight.w800),
         ),
         const SizedBox(height: 10),
@@ -29,7 +25,10 @@ class AddressSection extends StatelessWidget {
           children: [
             Expanded(
               child: TextFormField(
+                key: ValueKey(viewModel.address),
+                initialValue: viewModel.address,
                 readOnly: true,
+                maxLines: 1,
                 decoration: InputDecoration(
                   hintText: "주소를 입력해주세요",
                   hintStyle: TextStyle(
@@ -63,7 +62,19 @@ class AddressSection extends StatelessWidget {
             const SizedBox(width: 10),
 
             GestureDetector(
-              onTap: () {},
+              onTap: () async {
+                final selected = await showLocationDialog(context);
+                if (selected == null) return;
+
+                final geo = await viewModel.geocodeAddress(selected.address);
+                if (geo == null) return;
+
+                viewModel.setLocation(
+                  address: selected.address,
+                  latitude: geo['latitude']!.toString(),
+                  longitude: geo['longitude']!.toString(),
+                );
+              },
               child: Container(
                 padding: const EdgeInsets.all(14.0),
                 decoration: BoxDecoration(
@@ -75,7 +86,47 @@ class AddressSection extends StatelessWidget {
             ),
           ],
         ),
+
+        const SizedBox(height: 10),
+
+        if (viewModel.address != null) _NaverMap(),
       ],
+    );
+  }
+}
+
+class _NaverMap extends StatelessWidget {
+  const _NaverMap({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    final latLng = context.select<ScheduleAddViewModel, NLatLng?>((vm) {
+      if (vm.latitude == null || vm.longitude == null) return null;
+      return NLatLng(double.parse(vm.latitude!), double.parse(vm.longitude!));
+    });
+
+    if (latLng == null) {
+      return const SizedBox(
+        height: 200,
+        child: Center(child: Text("위치 정보 없음")),
+      );
+    }
+
+    return Container(
+      width: 400,
+      height: 400,
+      decoration: BoxDecoration(borderRadius: BorderRadius.circular(10.0)),
+      child: NaverMap(
+        options: NaverMapViewOptions(
+          initialCameraPosition: NCameraPosition(target: latLng, zoom: 15),
+        ),
+        onMapReady: (controller) {
+          controller.clearOverlays();
+          controller.addOverlay(
+            NMarker(id: "schedule_location", position: latLng),
+          );
+        },
+      ),
     );
   }
 }
