+++ lib/features/schedule/presentation/views/detail/widgets/emotion_color_section.dart
@@ -1,0 +1,34 @@
+import 'package:dutytable/features/schedule/presentation/viewmodels/schedule_detail_view_model.dart';
+import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
+
+class EmotionColorSection extends StatelessWidget {
+  /// 일정 상세 - 감정 및 색
+  const EmotionColorSection({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    final viewModel = context.watch<ScheduleDetailViewModel>();
+
+    return Padding(
+      padding: const EdgeInsets.symmetric(horizontal: 16.0),
+      child: Row(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          Text(viewModel.emotionTag ?? "", style: TextStyle(fontSize: 26)),
+
+          const SizedBox(width: 16),
+
+          Container(
+            width: 32,
+            height: 32,
+            decoration: BoxDecoration(
+              color: Color(int.parse(viewModel.colorValue)),
+              shape: BoxShape.circle,
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
