+++ lib/features/schedule/presentation/views/detail/widgets/memo_section.dart
@@ -1,0 +1,42 @@
+import 'package:dutytable/features/schedule/presentation/viewmodels/schedule_detail_view_model.dart';
+import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
+
+/// 일정 상세 - 메모
+class MemoSection extends StatelessWidget {
+  const MemoSection({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    final viewModel = context.watch<ScheduleDetailViewModel>();
+
+    return Padding(
+      padding: const EdgeInsets.symmetric(horizontal: 16.0),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          const Text(
+            "메모",
+            style: TextStyle(fontSize: 16, fontWeight: FontWeight.w800),
+          ),
+
+          const SizedBox(height: 10),
+
+          TextFormField(
+            initialValue: viewModel.memo,
+            maxLength: 300,
+            maxLines: 4,
+            enabled: false,
+            onChanged: (value) => viewModel.memo = value,
+            decoration: InputDecoration(
+              hintText: "메모를 입력하세요 (최대 300자)",
+              border: OutlineInputBorder(
+                borderRadius: BorderRadius.circular(10),
+              ),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
